requirements.txt
requests
beautifulsoup4
lxml
tqdm
sentence-transformers
faiss-cpu
numpy
streamlit